Make password errors fatal on connect.
If the user provides a password to PoCoCM->spawn(), it is sent to MPD on
connect. Previously, if the password was incorrect, it would yield an
'mpd_error' event in listening sessions for a command that the user had not
themselves sent and thus might not be prepared to handle. This behaviour is
changed so that instead an 'mpd_connect_error_fatal' event is raised and the
connection is terminated. It is assumed that the user set the password with
the intent to use it, and that that password being incorrect is an error
which they would want to catch.